+++ xdef/src/main/java/org/xdef/impl/parsers/XDParseCDATA.java
@@ -33,17 +33,24 @@ public class XDParseCDATA extends XDParserAbstract {
 		} else {
 			p.setEos();
 		}
-		Charset chset = xnode != null ? xnode.getXDPool().getLegalStringCharset() : null;
-		if (chset != null) {
-			byte[] bytes = s.getBytes(chset);
-			if (bytes.length != s.length()) {
-				p.errorWithString(XDEF.XDEF823, chset.name());
-			} else {
-				String t = new String(bytes, chset);
-				if (!s.equals(t)) {
-					p.errorWithString(XDEF.XDEF823, chset.name());
+		Charset[] chsets = xnode != null ? xnode.getXDPool().getLegalStringCharsets() : null;
+		if (chsets != null && chsets.length > 0) {
+			String err = "";
+			for (Charset chset : chsets) {
+				byte[] bytes = s.getBytes(chset);
+				if (bytes.length != s.length()) {
+					err += ' ' + chset.name();
+				} else {
+					String t = new String(bytes, chset);
+					if (!s.equals(t)) {
+						err += ' ' + chset.name();
+					} else {
+						return;
+					}
 				}
+				p.error(XDEF.XDEF823, err.trim());
 			}
+
 		}
 	}
 	@Override