+++ xdef/src/main/java/org/xdef/XDConstants.java
@@ -80,8 +80,8 @@ public interface XDConstants {
 	public static final String XDPROPERTY_MAXYEAR = "xdef_maxyear";
 	/** Property defines legal values of dates if year is out of range.*/
 	public static final String XDPROPERTY_SPECDATES = "xdef_specdates";
-	/** Property defines code table name for parsed string values (default is do not check).*/
-	public static final String XDPROPERTY_STRING_CODES = "xdef_string_code";
+	/** Property defines code table names for parsed string values (default is do not check).*/
+	public static final String XDPROPERTY_STRING_CODES = "xdef_string_codes";
 	/** Property defines default time zone.*/
 	public static final String XDPROPERTY_DEFAULTZONE = "xdef_defaultZone";